dealt with some warnings about implicit casts to float

+++ src/gfx/Sprite.cpp
@@ -15,7 +15,7 @@ namespace bustout
 			throw std::runtime_error("failed to load texture");
 
 		m_sprite.setTexture(m_texture);
-		m_sprite.setScale(1 / m_sprite.getTextureRect().width, 1 / m_sprite.getTextureRect().height);
+		m_sprite.setScale(1 / (float)m_sprite.getTextureRect().width, 1 / (float)m_sprite.getTextureRect().height);
 		m_sprite.setOrigin(m_sprite.getTextureRect().width * 0.5f, m_sprite.getTextureRect().height * 0.5f);
 	}
 
@@ -23,7 +23,7 @@ namespace bustout
 	{
 		m_sprite.setTextureRect(rect);
 		m_sprite.setOrigin(rect.width * 0.5f, rect.height * 0.5f);
-		m_sprite.setScale(1 / m_sprite.getTextureRect().width, 1 / m_sprite.getTextureRect().height);
+		m_sprite.setScale(1 / (float)m_sprite.getTextureRect().width, 1 / (float)m_sprite.getTextureRect().height);
 	}
 
 	void Sprite::setScale(const sf::Vector2f& scale)
